LPS-79388 (announcements-uad-test) Rename UADEntityTestHelper to UADTestHelper
