+++ modules/apps/collaboration/announcements/announcements-uad-test/src/testIntegration/java/com/liferay/announcements/uad/anonymizer/test/AnnouncementsEntryUADAnonymizerTest.java
@@ -17,7 +17,7 @@ package com.liferay.announcements.uad.anonymizer.test;
 import com.liferay.announcements.kernel.model.AnnouncementsEntry;
 import com.liferay.announcements.kernel.service.AnnouncementsEntryLocalService;
 import com.liferay.announcements.uad.constants.AnnouncementsUADConstants;
-import com.liferay.announcements.uad.test.AnnouncementsEntryUADEntityTestHelper;
+import com.liferay.announcements.uad.test.AnnouncementsEntryUADTestHelper;
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
@@ -59,8 +59,7 @@ public class AnnouncementsEntryUADAnonymizerTest
 		throws Exception {
 
 		AnnouncementsEntry announcementsEntry =
-			_announcementsEntryUADEntityTestHelper.addAnnouncementsEntry(
-				userId);
+			_announcementsEntryUADTestHelper.addAnnouncementsEntry(userId);
 
 		if (deleteAfterTestRun) {
 			_announcementsEntries.add(announcementsEntry);
@@ -115,8 +114,7 @@ public class AnnouncementsEntryUADAnonymizerTest
 	private AnnouncementsEntryLocalService _announcementsEntryLocalService;
 
 	@Inject
-	private AnnouncementsEntryUADEntityTestHelper
-		_announcementsEntryUADEntityTestHelper;
+	private AnnouncementsEntryUADTestHelper _announcementsEntryUADTestHelper;
 
 	@Inject(
 		filter = "model.class.name=" + AnnouncementsUADConstants.CLASS_NAME_ANNOUNCEMENTS_ENTRY
